JoinSplit partially implemented

+++ src/liblelantus/lelantus_prover.cpp
@@ -9,7 +9,7 @@ void LelantusProver::proof(
         const std::vector<std::vector<PublicCoin>>& anonymity_sets,
         const Scalar& Vin,
         const std::vector<std::pair<PrivateCoin, uint32_t>>& Cin,
-        const std::vector<uint64_t>& indexes,
+        const std::vector<size_t>& indexes,
         const Scalar& Vout,
         const std::vector<PrivateCoin>& Cout,
         const Scalar& f,
@@ -54,7 +54,7 @@ void LelantusProver::proof(
 void LelantusProver::generate_sigma_proofs(
         const std::vector<std::vector<PublicCoin>>& c,
         const std::vector<std::pair<PrivateCoin, uint32_t>>& Cin,
-        const std::vector<uint64_t>& indexes,
+        const std::vector<size_t>& indexes,
         Scalar& x,
         std::vector<Scalar>& Yk_sum,
         std::vector<SigmaPlusProof<Scalar, GroupElement>>& sigma_proofs){
