+++ src/liblelantus/lelantus_prover.h
@@ -15,7 +15,7 @@ public:
             const std::vector<std::vector<PublicCoin>>& anonymity_sets,
             const Scalar& Vin,
             const std::vector<std::pair<PrivateCoin, uint32_t>>& Cin,
-            const std::vector <uint64_t>& indexes,
+            const std::vector <size_t>& indexes,
             const Scalar& Vout,
             const std::vector <PrivateCoin>& Cout,
             const Scalar& f,
@@ -25,7 +25,7 @@ private:
     void generate_sigma_proofs(
             const std::vector<std::vector<PublicCoin>>& c,
             const std::vector<std::pair<PrivateCoin, uint32_t>>& Cin,
-            const std::vector<uint64_t>& indexes,
+            const std::vector<size_t>& indexes,
             Scalar& x,
             std::vector<Scalar>& Yk_sum,
             std::vector<SigmaPlusProof<Scalar, GroupElement>>& sigma_proofs);
